+++ CHANGELOG.md
@@ -3,6 +3,13 @@ Upgraders: for a fresh start, consider running these commands
 * `git clean -xdf` 
 * `npm install`
 
+<a name="0.2.18"></a>
+# 0.2.18 (2016-11-30)
+* remove `exclude` clause from `tsconfig.json`; it was just confusing people
+* karma.config + karma-test-shim can handle multiple spec source paths (issue #294)
+* cosmetic `app.component.spec.ts` changes
+* cosmetic `karma.config.js` changes
+
 <a name="0.2.17"></a>
 # 0.2.17 (2016-11-16)
 * Conform to updated QuickStart advice